Fix corrupt locations being forgotten in prisoner v3 to v4 migration
If we set unknown location to true here, then it will just look like
an unknown location, and the fact that the location was corrupt will
be lost. Instead, we need to *not* mark it as unknown, then code in
PrisonerDataHandler will notice the lack of a location while a location
is claimed to be known, a situation that means that the location got
corrupt.