Resolve worlds from ImmutableLocations lazily/when required (#31)
Store the world name instead of the World object, and resolve the World object when required (when converting to a mutable Location). In the plugin that is just when a prisoner needs to be teleported, but plugins interacting with the API can trigger that as well.
This commit also migrates prisoner and jail data from storing Location objects to ImmutableLocation ones. The problem this solves remains for when the data upgraders run the first time, as it needs the Location objects to convert to ImmutableLocation and data upgraders run eagerly as all data is loaded as soon as the server starts ticking, so the problem of "world may not be loaded yet" is present there as well, but that won't be a problem in the future until the world is actually needed again